Implement Add Category Fixes #17
