+++ src/mocks/handlers/category.handler.ts
@@ -2,9 +2,9 @@ import { rest } from 'msw'
 import _ from 'lodash'
 import { faker } from '@faker-js/faker'
 import { CreatePagedResponse } from '../handlers.utility'
-import type { Category } from '~/models/Category'
+import type { Category, CategoryCreateModel } from '~/models/Category'
 
-const categories = _.times(77, createFakeCategory)
+const categories = _.times(7, createFakeCategory)
 const handlers = [
   rest.get('/api/Category', (req, res, ctx) => {
     const response = CreatePagedResponse<Category>(req, categories)
@@ -14,6 +14,22 @@ const handlers = [
       ctx.json(response),
     )
   }),
+  rest.post('/api/Category', async (req, res, ctx) => {
+    const newItem = await req.json<CategoryCreateModel>()
+    const category: Category = {
+      id: faker.datatype.number({ max: 2000 }),
+      name: newItem.name,
+      productsCount: 0,
+      subItems: [],
+    }
+    categories.push(category)
+    return res(
+      ctx.status(200),
+      ctx.delay(200),
+      ctx.json(category),
+    )
+  }),
+
 ]
 
 function createFakeCategory(): Category {
