+++ src/styles/main.scss
@@ -55,7 +55,6 @@ body {
 
 .page-title {
   font-size: 1.1rem;
-  padding-bottom: 1.2rem;
 }
 @import 'dark';
 @import 'utils/override';
